// src/screens/editAccount/editAccount.types.ts
export type FormData = {
  social_name: string;
  surname: string;
  zip_code: string;
  street: string;
  number: string;
  neighborhood: string;
  city: string;
  state: string;
  phone: string;
  date_of_birth: string;
  gender: string;
};

export type DataProps = {
  text: string;
  field: string;
};